refactor(common): split extra data modules
